requirements.txt
Pillow
fuzzywuzzy
python-Levenshtein
matplotlib
requests
beautifulsoup4
selenium